Add files via upload
In this article you can see the relevant scientific research I did for the theoretical implementation of the application. You will also see the analysis and parts of the code for the most important functions of the application. Finally you will see the functions running on IOS operating system. This article is my thesis for the completion of my studies at the department of digital systems at the university of Piraeus.